requirements.txt
streamlit>=1.36
openai>=1.40
python-dotenv>=1.0.1
pandas>=2.2
numpy>=1.26
faiss-cpu>=1.8.0.post1